feat: adding mdns service discovery (#102)
* feat: adding in initial checks for mdns

* updating mdns and server setup to include mdns

* adding ifaces cmd

* updating todo list

* fix: solving linting issues

* adding more log messages

* adding docker compose file

* fixing up pipelines

* adding more summaries

* Revert "adding more summaries"

This reverts commit 83e90d6fb670e7e088e5b9bd481acb0003b3fdef.

+++ server/internal/grpc/config.go
@@ -5,11 +5,11 @@ import "github.com/DaanV2/mechanus/server/pkg/config"
 var (
 	APIConfig = config.New("api")
 	HostFlag  = APIConfig.String("api.host", "", "What host to bind on, such as: \"\", \"localhost\" or \"0.0.0.0\"")
-	PortFlag  = APIConfig.Int("api.port", 8666, "The port to server api traffic to")
+	PortFlag  = APIConfig.UInt16("api.port", 8666, "The port to server api traffic to")
 )
 
 type Config struct {
-	Port int
+	Port uint16
 	Host string
 }
